+++ src/package/vendors/text/CodeViewer.vue
@@ -59,6 +59,7 @@ const languageMap: Record<string, string> = {
   ts: 'typescript',
   tsx: 'typescript',
   txt: 'plaintext',
+  umd: 'javascript',
   vue: 'xml',
   xml: 'xml',
   yaml: 'yaml',